feat: inicio de creación de la factura

+++ crud/views.py
@@ -1,8 +1,9 @@
-from django.shortcuts import render
-from django.shortcuts import redirect, get_object_or_404
-
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Producto, Categoria, Marca
-from .forms import ProductoForm, CategoriaForm, MarcaForm
+from .forms import ProductoForm, CategoriaForm, MarcaForm, FacturaForm, DetalleFacturaForm
+
+from django.http import HttpResponseNotAllowed, JsonResponse
+from django.core.paginator import Paginator
 
 # Create your views here.
 def index(request):
@@ -31,10 +32,12 @@ def editar(request, id):
         precioVenta = request.POST.get('precioVenta')
         precioCompra = request.POST.get('precioCompra')
         estado = request.POST.get('estado')
+        stock = request.POST.get('stock')
         producto.nombre = nombre
         producto.precioVenta = precioVenta
         producto.precioCompra = precioCompra
         producto.estado = estado
+        producto.stock = stock
         producto.save()
         return redirect('crud:verProducto')
     return render(request, 'editarProducto.html', {'producto': producto})
@@ -57,7 +60,8 @@ def crearCategoria(request):
     if request.method == 'POST':
         form = CategoriaForm(request.POST)
         if form.is_valid():
-            form.save() 
+            form.save()
+            return redirect('crud:crearCategoria')
     else:
         form = CategoriaForm()
     
@@ -69,14 +73,17 @@ def editarCategoria(request, id):
         nombre = request.POST.get('nombre')
         categoria.nombre = nombre
         categoria.save()
-        return redirect('index')
+        return redirect('crud:verCategoria')
     return render(request, 'editarCategoria.html', {'categoria': categoria})
 
 def eliminarCategoria(request, id):
-    categoria = Categoria.objects.get(id=id)
-    categoria.delete()
-    return redirect('index')
+    categoria = get_object_or_404(Categoria, id=id)
 
+    if request.method == 'POST':
+        categoria.delete()
+        return redirect('crud:verCategoria')   
+    return render(request, 'confirmarEliminarCategoria.html', {'categoria': categoria})
+    
 def verCategoria(request):
     categorias = Categoria.objects.all()
     return render(request, 'verCategoria.html', {'categorias': categorias})
@@ -88,6 +95,7 @@ def crearMarca(request):
         form = MarcaForm(request.POST)
         if form.is_valid():
             form.save()
+            return redirect('crud:crearMarca')
     else:
         form = MarcaForm()
     
@@ -101,18 +109,61 @@ def editarMarca(request, id):
         marca.nombre = nombre
         marca.estado = estado
         marca.save()
-        return redirect('index')
+        return redirect('crud:verMarca')
     return render(request, 'editarMarca.html', {'marca': marca})
 
 def eliminarMarca(request, id):
-    marca = Marca.objects.get(id=id)
-    marca.delete()
-    return redirect('index')
+    marca = get_object_or_404(Marca, id=id)
+    
+    if request.method == 'POST':
+        marca.delete()
+        return redirect('crud:verMarca')
+    return render(request, 'confirmarEliminarMarca.html', {'marca': marca})
 
 def verMarca(request):
     marcas = Marca.objects.all()
     return render(request, 'verMarca.html', {'marcas': marcas})
 
 
+#FACTURAS
+def crearFactura(request):
+    if request.method == 'POST':
+        form = FacturaForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('crud:crearFactura')
+    else:
+        form = FacturaForm()
+    
+    return render(request, 'crearFactura.html', {'form': form})
 
+def crearDetalleFactura(request):
+    if request.method == 'POST':
+        form = DetalleFacturaForm(request.POST)
+        if form.is_valid():
+            form.save()
+            return redirect('crud:crearFactura')
+    else:
+        form = DetalleFacturaForm()
+    
+    return render(request, 'crearDetalleFactura.html', {'form': form})
+
+def buscar_productos(request):
+    query = request.GET.get('q', '')
+    productos = Producto.objects.filter(nombre__icontains=query) if query else Producto.objects.all()
+    productos_data = [{'id': p.id, 'nombre': p.nombre, 'precio':  str( p.precioVenta)} for p in productos]
+    print(productos_data)
+    return render(request, 'buscar_productos.html', {'productos': productos, 'query': query})
+
+def detalle_producto(request, id):
+    producto = get_object_or_404(Producto, id=id)
+    return render(request, 'productos/detalle_producto.html', {'producto': producto})
+
+
+def agregar_carrito(request, id):
+    # Aquí puedes agregar la lógica para añadir el producto al carrito de compras
+    producto = get_object_or_404(Producto, id=id)
+    # Supongamos que tienes un método para agregar al carrito
+    agregar_producto_al_carrito(request, producto)
+    return JsonResponse({'message': 'Producto agregado al carrito'})
